refactor: split policyreport api files
